refactor: restructure library implementations
