+++ libs/mql-interpreter/src/index.ts
@@ -2,4 +2,4 @@ export { Parser } from "./parser/parser";
 export { semanticCheck } from "./semantic/checker";
 export { builtinSignatures } from "./libs/signatures";
 export { Runtime } from "./runtime/runtime";
-export { createBacktestLibs } from "./libs/backtest";
+export { createBacktestLibs, createLiveLibs, BrokerApi } from "./libs";
