--- libs/mql-interpreter/src/libs/builtins/OrderSend.ts
@@ -1,3 +1,0 @@
-import type { BuiltinFunction } from "./types";
-
-export const OrderSend: BuiltinFunction = () => 0;
