+++ libs/mql-interpreter/src/libs/common/iATR.ts
