+++ libs/mql-interpreter/src/libs/common/iMACD.ts
