+++ libs/mql-interpreter/src/libs/common/indicator.ts
