+++ libs/mql-interpreter/src/libs/common/math.ts
