+++ libs/mql-interpreter/src/libs/common/strings.ts
