+++ libs/mql-interpreter/src/libs/domain/marketInformation/live.ts
@@ -1,0 +1,10 @@
+import type { BuiltinFunction } from "../../common/types";
+
+export function createMarketInformation(): Record<string, BuiltinFunction> {
+  return {
+    MarketInfo: () => 0,
+    SymbolsTotal: () => 0,
+    SymbolName: () => "",
+    SymbolSelect: () => true,
+  };
+}
