+++ libs/mql-interpreter/src/libs/domain/trading/live.ts
@@ -1,0 +1,11 @@
+import type { BuiltinFunction } from "../../common/types";
+
+export interface TradingApi {
+  sendOrder: (request: any) => any;
+}
+
+export function createTrading(api: TradingApi): Record<string, BuiltinFunction> {
+  return {
+    OrderSend: api.sendOrder.bind(api),
+  };
+}
