+++ libs/mql-interpreter/test/runtime/builtins/strings.test.ts
@@ -16,7 +16,7 @@ import {
   StringSplit,
   StringToLower,
   StringToUpper,
-} from "../../../src/libs/builtins/strings";
+} from "../../../src/libs/common/strings";
 import { describe, it, expect } from "vitest";
 
 describe("string builtins", () => {
